cutting out another scheduling overhead by making PRELOAD special to assigned bufs.
```
~/code/tinygrad (master) > SCHEDULE_ONLY=1 python3 test/external/external_benchmark_schedule.py
***** model tensor in     62.26 ms
***** model schedule in  205.19 ms
all 267.53 ms

~/code/tinygrad (assign_adjacents) > SCHEDULE_ONLY=1 python3 test/external/external_benchmark_schedule.py
***** model tensor in     62.27 ms
***** model schedule in  195.73 ms
all 258.07 ms
```